Korretur Transaction Klasse 2

+++ src/main/java/entityclasses/Transaction.java
@@ -19,11 +19,11 @@ public class Transaction {
         this.date = date;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
